@@ -6,11 +6,16 @@
   "author": "",
   "license": "ISC",
   "devDependencies": {
+    "babel-eslint": "^7.1.1",
     "babel-preset-es2015": "^6.1.18",
     "babel-preset-react": "^6.1.18",
     "babelify": "^7.2.0",
     "browserify": "^12.0.1",
-    "eslint": "^2.11.0",
+    "eslint": "^3.13.1",
+    "eslint-config-uber-es2015": "^3.0.1",
+    "eslint-config-uber-jsx": "^3.0.1",
+    "eslint-plugin-babel": "^4.0.0",
+    "eslint-plugin-react": "^6.10.0",
     "gulp": "^3.9.0",
     "gulp-live-server": "0.0.29",
     "gulp-sass": "^2.1.0",
